+++ app/controllers/api/users/tasks_controller.rb
@@ -17,11 +17,16 @@ private
 
   def create_task_params
     parameters = fetch_resource_params(:task, [:label], [:planned_at, :project_id])
-    parameters[:state] = 'started'
-    parameters[:started_at] = DateTime.now
+    project = Project.find(parameters[:project_id]) if parameters.has_key?(:project_id)
     if parameters.has_key?(:planned_at)
       parameters[:state] = 'planned'
       parameters[:planned_at] = parameters[:planned_at].to_datetime
+      parameters[:started_at] = DateTime.now
+    elsif project.nil? || project.started?
+      parameters[:state] = 'started'
+      parameters[:started_at] = DateTime.now
+    else
+      parameters[:state] = 'newed'
     end
     parameters
   end